+++ FFXIVAPP.Client/Network/ServerConnection.cs
@@ -43,11 +43,11 @@ namespace FFXIVAPP.Client.Network
         public override bool Equals(object obj)
         {
             var connection = obj as ServerConnection;
-            if (connection == null || (int) SourceAddress != (int) connection.SourceAddress || ((int) DestinationAddress != (int) connection.DestinationAddress || (int) SourcePort != (int) connection.SourcePort))
+            if (connection == null || (int) SourceAddress != (int) connection.SourceAddress || ((int) DestinationAddress != (int) connection.DestinationAddress || SourcePort != connection.SourcePort))
             {
                 return false;
             }
-            return (int) DestinationPort == (int) connection.DestinationPort;
+            return DestinationPort == connection.DestinationPort;
         }
 
         public override int GetHashCode()
